Change release_images.archive_org_code to bigint

+++ db/migrate/20201111194756_create_release_images.rb
@@ -5,7 +5,7 @@ class CreateReleaseImages < ActiveRecord::Migration[6.0]
       t.boolean :back, default: false, null: false
       t.string :note
       t.references :release, null: false, foreign_key: true
-      t.integer :archive_org_code
+      t.bigint :archive_org_code
 
       t.timestamps
     end
